+++ js/data.js
@@ -1,0 +1,25 @@
+const DESCRIPTIONS = [
+    'Lovely day',
+    'Nice weather',
+    'Working on project',
+    'Outdoor weekend',
+    'Everything in its right place',
+    'Hanging out with family',
+    'My new car',
+    'Enjoying the sun',
+    'I love Italy',
+    '*Some random description*'
+  ];
+  
+  const NAMES = ['Georgy', 'Sacha', 'Max', 'Benj', 'Josh', 'Drew', 'Callum', 'Toby'];
+  
+  const MESSAGES = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+export {DESCRIPTIONS, NAMES, MESSAGES};
